refactor(161): add index.ts for exceptions folder
