+++ src/common/exceptions/index.ts
@@ -1,0 +1,2 @@
+export * from './base-api.exception'
+export * from './error-codes'
